logistf.control {logistf}	R Documentation

Control Parameters for logistf

Description

Sets parameters for Newton-Raphson iteration in Firth's penalized-likelihood logistic regression

Usage

logistf.control(maxit = 25, maxhs = 5, maxstep = 5, lconv = 1e-05, gconv = 1e-05, 
    xconv = 1e-05, collapse=TRUE)

Arguments

maxit
the maximum number of iterations

maxhs
the maximum number of step-halvings in one iteration. The increment of the beta vector within one iteration is divided by 2 if the new beta leads to a decrease in log likelihood.

maxstep
specifies the maximum step size in the beta vector within one iteration.

lconv
specifies the convergence criterion for the log likelihood.

gconv
specifies the convergence criterion for the first derivative of the log likelihood (the score vector).

xconv
specifies the convergence criterion for the parameter estimates.

collapse
if TRUE, evaluates all unique combinations of x and y and collapses data set. This may save computing time with large data sets with only categorical (binary) covariates.

Details

logistf.control() is used by logistf and logistftest to set control parameters to default values. Different values can be specified, e. g., by logistf(..., control= logistf.control(maxstep=1)).

Value

maxit
the maximum number of iterations

maxhs
the maximum number of step-halvings in one iteration. The increment of the beta vector within one iteration is divided by 2 if the new beta leads to a decrease in log likelihood.

maxstep
specifies the maximum step size in the beta vector within one iteration.

lconv
specifies the convergence criterion for the log likelihood.

gconv
specifies the convergence criterion for the first derivative of the log likelihood (the score vector).

xconv
specifies the convergence criterion for the parameter estimates.

collapse
if TRUE, evaluates all unique combinations of x and y and collapses data set.

Author(s)

Georg Heinze

Examples

data(sexagg)
fit2<-logistf(case ~ age+oc+vic+vicl+vis+dia, data=sexagg, weights=COUNT, 
   control=logistf.control(maxstep=1))
summary(fit2)

[Package logistf version 1.22 Index]