logistpl.control {logistf}	R Documentation

Control Parameters for logistf Profile Likelihood Confidence Interval Estimation

Description

Sets parameters for modified Newton-Raphson iteration for finding profile likelihood confidence intervals in Firth's penalized likelihood logistic regression

Usage

logistpl.control(maxit=100, maxhs=5, maxstep=5, lconv=0.00001, xconv=0.00001, 
   ortho=FALSE, pr=FALSE)

Arguments

maxit
the maximum number of iterations

maxhs
the maximum number of step-halvings in one iteration. The increment of the beta vector within one iteration is divided by 2 if the new beta leads to a decrease in log likelihood.

maxstep
specifies the maximum step size in the beta vector within one iteration.

lconv
specifies the convergence criterion for the log likelihood.

xconv
specifies the convergence criterion for the parameter estimates.

ortho
requests orthogonalization of variable for which confidence intervals are computed with respect to other covariates.

pr
request rotation of the matrix spanned by the covariates

Details

logistpl.control() is used by logistf to set control parameters to default values when computing profile likelihood confidence intervals. Different values can be specified, e. g., by logistf(..., control= logistf.control(maxstep=1)).

Value

maxit
the maximum number of iterations

maxhs
the maximum number of step-halvings in one iteration. The increment of the beta vector within one iteration is divided by 2 if the new beta leads to a decrease in log likelihood.

maxstep
specifies the maximum step size in the beta vector within one iteration.

lconv
specifies the convergence criterion for the log likelihood.

xconv
specifies the convergence criterion for the parameter estimates.

ortho
specifies if orthogonalization is requested.

pr
specifies if rotation is requested

Author(s)

Georg Heinze

Examples

data(sexagg)
fit2<-logistf(case ~ age+oc+vic+vicl+vis+dia, data=sexagg, weights=COUNT, 
   plcontrol=logistpl.control(maxstep=1))
summary(fit2)

[Package logistf version 1.22 Index]