predict.flic {logistf}	R Documentation

Predict Method for flic Fits

Description

Obtains predictions from a fitted flic object.

Usage

## S3 method for class 'flic'
predict(
  object,
  newdata,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  ...
)

Arguments

object
A fitted object of class flic.

newdata
Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.

type
The type of prediction required. The default is on the scale of the linear predictors. The alternative response gives the predicted probabilities. Type terms returns a matrix with the fitted values of each term in the formula on the linear predictor scale.

se.fit
If TRUE(default = FALSE) standard errors are computed.

...
further arguments passed to or from other methods.

Details

If newdata is omitted the predictions are based on the data used for the fit.

Value

A vector or matrix of predictions

[Package logistf version 1.24.1 Index]