predict.logistf {logistf}	R Documentation

Predict Method for logistf Fits

Description

Obtains predictions from a fitted logistf object.

Usage

## S3 method for class 'logistf'
predict(
  object,
  newdata,
  type = c("link", "response", "terms"),
  flic = FALSE,
  se.fit = FALSE,
  reference,
  na.action = na.pass,
  ...
)

Arguments

object
A fitted object of class logistf.

newdata
Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.

type
The type of prediction required. The default is on the scale of the linear predictors. The alternative response gives the predicted probabilities. Type terms returns a matrix with the fitted values of each term in the formula on the linear predictor scale.

flic
If TRUE(default = FALSE), predictions are computed with intercept correction.

se.fit
If TRUE(default = FALSE) standard errors are computed.

reference
A named vector of reference values for each variable for type="terms".

na.action
Function determining what should be done with missing values in newdata. The default is to predict NA.

...
further arguments passed to or from other methods.

Details

If newdata is omitted the predictions are based on the data used for the fit.

Value

A vector or matrix of predictions.

[Package logistf version 1.24.1 Index]