profile.logistf {logistf}	R Documentation

Compute Profile Penalized Likelihood

Description

Evaluates the profile penalized likelihood of a variable based on a logistf model fit.

Usage

## S3 method for class 'logistf'
profile(fitted, which, variable, steps = 100, pitch = 0.05, limits, 
   alpha = 0.05, firth = TRUE, legends = TRUE, control, plcontrol, plot = FALSE,
   ...)

Arguments

fitted
An object fitted by logistf

which
A righthand formula to specify the variable for which the profile should be evaluated, e.g., which=~X).

variable
Alternatively to which, a variable name can be given, e.g., variable="X"

steps
number of steps in evaluating the profile likelihood

pitch
alternatively to steps, one may specify the step width in multiples of standard errors

limits
lower and upper limits of parameter values at which profile likelihood is to be evaluated

alpha
the significance level (1-α the confidence level, 0.05 as default).

firth
use of Firth's penalized maximum likelihood (firth=TRUE, default) or the standard maximum likelihood method (firth=FALSE) for the logistic regression.

legends
legends to be included in the optional plot

control
Controls Newton-Raphson iteration. Default is control= logistf.control(maxstep, maxit, maxhs, lconv, gconv, xconv)

plcontrol
Controls Newton-Raphson iteration for the estimation of the profile likelihood confidence intervals. Default is plcontrol= logistpl.control(maxstep, maxit, maxhs, lconv, xconv, ortho, pr)

plot
If TRUE, profile likelihood is plotted. This parameter becomes obsolete as a generic plot function is now provided.

...
Further arguments to be passed.

Value

An object of class logistf.profile with the following items:

beta
parameter values at which likelihood was evaluated

stdbeta
parameter values divided by standard error

profile
profile likelihood, standardized to 0 at maximum of likelihood. The values in profile are given as minus χ^2.

loglike
unstandardized profile likelihood

signed.root
signed root (z) of χ^2 values (negative for values below the maximum likelihood estimate, positive for values above the maximum likelihood estimate)

cdf
profile likelihood expressed as cumulative distribution function, obtained as Φ(z), where Φ denotes the standard normal distribution function.

Author(s)

Georg Heinze and Meinhard Ploner

See Also

plot.profile.logistf

Examples

data(sex2)
fit<-logistf(case ~ age+oc+vic+vicl+vis+dia, data=sex2)
plot(profile(fit,variable="dia"))
plot(profile(fit,variable="dia"), "cdf")
plot(profile(fit,variable="dia"), "density")

[Package logistf version 1.22 Index]