logistf.control {logistf}	R Documentation

Control Parameters for logistf

Description

Sets parameters for iterations in Firth's penalized-likelihood logistic regression.

Usage

logistf.control(
  maxit = 25,
  maxhs = 0,
  maxstep = 5,
  lconv = 1e-05,
  gconv = 1e-05,
  xconv = 1e-05,
  collapse = TRUE,
  fit = "NR"
)

Arguments

maxit
The maximum number of iterations

maxhs
The maximum number of step-halvings in one iteration. The increment of the beta vector within one iteration is divided by 2 if the new beta leads to a decrease in log likelihood.

maxstep
Specifies the maximum step size in the beta vector within one iteration. Set to -1 for infinite stepsize.

lconv
Specifies the convergence criterion for the log likelihood.

gconv
Specifies the convergence criterion for the first derivative of the log likelihood (the score vector).

xconv
Specifies the convergence criterion for the parameter estimates.

collapse
If TRUE, evaluates all unique combinations of x and y and collapses data set.

fit
Fitting method used. One of Newton-Raphson: "NR" or Iteratively reweighted least squares: "IRLS"

Details

logistf.control() is used by logistf and logistftest to set control parameters to default values. Different values can be specified, e. g., by logistf(..., control= logistf.control(maxstep=1)).

Value

maxit
The maximum number of iterations

maxhs
The maximum number of step-halvings in one iteration. The increment of the beta vector within one iteration is divided by 2 if the new beta leads to a decrease in log likelihood.

maxstep
Specifies the maximum step size in the beta vector within one iteration.

lconv
Specifies the convergence criterion for the log likelihood.

gconv
Specifies the convergence criterion for the first derivative of the log likelihood (the score vector).

xconv
Specifies the convergence criterion for the parameter estimates.

collapse
If TRUE, evaluates all unique combinations of x and y and collapses data set.

fit
Fitting method used. One of Newton-Raphson: "NR" or Iteratively reweighted least squares: "IRLS"

call
The function call.

Examples

data(sexagg)
fit2<-logistf(case ~ age+oc+vic+vicl+vis+dia, data=sexagg, weights=COUNT, 
control=logistf.control(maxstep=1))
summary(fit2)

[Package logistf version 1.24.1 Index]